Fix some -Wsign-conversion
In some cases, return value of send_request was not used.
In other cases, make implicit cast explicit. Casting
from signed from to unsigned is safe here as negative
return values have been checked.

+++ src/ping_thread.c
@@ -199,7 +199,7 @@ ping(void)
 				done = 1;
 			}
 			else {
-				totalbytes += numbytes;
+				totalbytes += (size_t) numbytes;
 				debug(LOG_DEBUG, "Read %d bytes, total now %d", numbytes, totalbytes);
 			}
 		}
